requirements.txt
Flask==2.3.2
flask-cors==3.0.9
werkzeug==3.0.1
markupsafe==2.1.1
bcrypt==4.0.1
gunicorn==20.1.0
sendgrid==6.9.7
python-dateutil
python-dotenv
google-cloud-firestore==2.11.1
google-cloud-logging==3.10
google-cloud-scheduler==2.13.4
google-cloud-tasks==2.16.4
google-cloud-storage
openai==1.35.13
beautifulsoup4==4.12.3